LPS-146878 Move graphQL queries to respective their file
